# backend/openapi.json
{
    "openapi": "3.1.1",
    "info": {
        "title": "Nook API",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:5999/"
        }
    ],
    "paths": {
...
        "/api/tags": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "operationId": "ListTags",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TagDto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tags"
                ],
                "operationId": "CreateTag",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTagRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TagDto"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tags/{id}": {
            "patch": {
                "tags": [
                    "Tags"
                ],
                "operationId": "PatchTag",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchTagRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TagDto"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tags"
                ],
                "operationId": "DeleteTag",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/tags/{id}/nodes": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "operationId": "ListTagNodes",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "pattern": "^-?(?:0|[1-9]\\d*)$",
                            "type": [
                                "integer",
                                "string"
                            ],
                            "format": "int32"
                        }
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TagNodesPage"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/nodes/{id}/tags": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "description": "Union of manual and inline (#hashtag) tags; each entry carries its source.",
                "operationId": "GetNodeTags",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TagDto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Tags"
                ],
                "description": "Replaces the manual tag set; inline tags collected from the document text are untouched.",
                "operationId": "SetNodeTags",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SetNodeTagsRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TagDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateTagRequest": {
                "required": [
                    "name",
                    "color"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "color": {
                        "type": [
                            "null",
                            "string"
                        ]
                    }
                }
            },
...
            "NodeCover": {
                "required": [
                    "type",
                    "value"
                ],
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string"
                    },
                    "value": {
                        "type": "string"
                    },
                    "position": {
                        "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
                        "type": [
                            "null",
                            "number",
                            "string"
                        ],
                        "format": "double"
                    }
                }
            },
            "NodeDto": {
                "required": [
                    "id",
                    "workspaceId",
                    "parentId",
                    "kind",
                    "title",
                    "icon",
                    "cover",
                    "position",
                    "pageSettings",
                    "archivedAt",
                    "deletedAt",
                    "createdAt",
                    "updatedAt",
                    "effectiveRole"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "workspaceId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "parentId": {
                        "type": [
                            "null",
                            "string"
                        ],
                        "format": "uuid"
                    },
                    "kind": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "icon": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/NodeIcon"
                            }
                        ]
                    },
                    "cover": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/NodeCover"
                            }
                        ]
                    },
                    "position": {
                        "type": "string"
                    },
                    "pageSettings": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/PageSettings"
                            }
                        ]
                    },
                    "archivedAt": {
                        "type": [
                            "null",
                            "string"
                        ],
                        "format": "date-time"
                    },
                    "deletedAt": {
                        "type": [
                            "null",
                            "string"
                        ],
                        "format": "date-time"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "effectiveRole": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "hasChildren": {
                        "type": "boolean",
                        "default": false
                    },
                    "properties": {}
                }
            },
            "NodeIcon": {
                "required": [
                    "type",
                    "value"
                ],
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string"
                    },
                    "value": {
                        "type": "string"
                    }
                }
            },
...
            "PageSettings": {
                "type": "object",
                "properties": {
                    "font": {
                        "type": "string",
                        "default": "default"
                    },
                    "smallText": {
                        "type": "boolean",
                        "default": false
                    },
                    "fullWidth": {
                        "type": "boolean",
                        "default": false
                    },
                    "locked": {
                        "type": "boolean",
                        "default": false
                    }
                }
            },
...
            "PatchTagRequest": {
                "required": [
                    "name",
                    "color"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "color": {
                        "type": [
                            "null",
                            "string"
                        ]
                    }
                }
            },
            "ProblemDetails": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "title": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "status": {
                        "pattern": "^-?(?:0|[1-9]\\d*)$",
                        "type": [
                            "null",
                            "integer",
                            "string"
                        ],
                        "format": "int32"
                    },
                    "detail": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "instance": {
                        "type": [
                            "null",
                            "string"
                        ]
                    }
                }
            },
...
            "SetNodeTagsRequest": {
                "required": [
                    "tagIds",
                    "names"
                ],
                "type": "object",
                "properties": {
                    "tagIds": {
                        "type": [
                            "null",
                            "array"
                        ],
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    "names": {
                        "type": [
                            "null",
                            "array"
                        ],
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "TagDto": {
                "required": [
                    "id",
                    "name",
                    "color",
                    "count"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "color": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "count": {
                        "pattern": "^-?(?:0|[1-9]\\d*)$",
                        "type": [
                            "integer",
                            "string"
                        ],
                        "format": "int32"
                    },
                    "source": {
                        "type": [
                            "null",
                            "string"
                        ]
                    }
                }
            },
            "TagNodesPage": {
                "required": [
                    "items",
                    "nextCursor"
                ],
                "type": "object",
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/NodeDto"
                        }
                    },
                    "nextCursor": {
                        "type": [
                            "null",
                            "string"
                        ]
                    }
                }
            },
...
        }
    },
...
}